zeroshot_code/api_based_models/requirements.txt
pandas
python-dotenv
google-generativeai
openai
xai-sdk
